Refactor quiz package to model and QuizModel to model.modelManager subpackage
