+++ src/main/java/seedu/address/MainApp.java
@@ -19,10 +19,10 @@ import seedu.address.logic.LogicManager;
 import seedu.address.model.Lessons;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.UserPrefs;
+import seedu.address.model.modelmanager.QuizModel;
+import seedu.address.model.modelmanager.QuizModelManager;
 import seedu.address.model.modelmanager.management.ManagementModel;
 import seedu.address.model.modelmanager.management.ManagementModelManager;
-import seedu.address.model.modelmanager.quiz.QuizModel;
-import seedu.address.model.modelmanager.quiz.QuizModelManager;
 import seedu.address.model.user.User;
 import seedu.address.storage.CsvLessonsStorage;
 import seedu.address.storage.CsvUserStorage;
