+++ src/main/java/seedu/address/logic/Logic.java
@@ -6,7 +6,7 @@ import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.modelmanager.quiz.QuizUiDisplayFormatter;
+import seedu.address.model.quiz.QuizUiDisplayFormatter;
 
 /**
  * API of the Logic component
