+++ src/main/java/seedu/address/logic/commands/quiz/QuizDifficultCommand.java
@@ -7,8 +7,8 @@ import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.modelmanager.Model;
-import seedu.address.model.modelmanager.quiz.QuizCard;
-import seedu.address.model.modelmanager.quiz.QuizModel;
+import seedu.address.model.modelmanager.QuizModel;
+import seedu.address.model.quiz.QuizCard;
 
 /**
  * Label a card as difficult or not
