+++ src/main/java/seedu/address/logic/commands/quiz/QuizStatusCommand.java
@@ -7,7 +7,7 @@ import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.modelmanager.Model;
-import seedu.address.model.modelmanager.quiz.QuizModel;
+import seedu.address.model.modelmanager.QuizModel;
 
 /**
  * Display the total attempts, streak and progress of the current session
