+++ src/main/java/seedu/address/model/session/Session.java
@@ -3,8 +3,8 @@ package seedu.address.model.session;
 import java.util.ArrayList;
 import java.util.List;
 
-import seedu.address.model.modelmanager.quiz.Quiz;
-import seedu.address.model.modelmanager.quiz.QuizCard;
+import seedu.address.model.quiz.Quiz;
+import seedu.address.model.quiz.QuizCard;
 import seedu.address.model.srscard.SrsCard;
