+++ src/test/java/seedu/address/logic/commands/management/DeleteLessonCommandTest.java
@@ -21,11 +21,11 @@ import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Lessons;
 import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.lesson.Lesson;
+import seedu.address.model.modelmanager.QuizModel;
 import seedu.address.model.modelmanager.management.ManagementModel;
-import seedu.address.model.modelmanager.quiz.Quiz;
-import seedu.address.model.modelmanager.quiz.QuizCard;
-import seedu.address.model.modelmanager.quiz.QuizModel;
-import seedu.address.model.modelmanager.quiz.QuizUiDisplayFormatter;
+import seedu.address.model.quiz.Quiz;
+import seedu.address.model.quiz.QuizCard;
+import seedu.address.model.quiz.QuizUiDisplayFormatter;
 import seedu.address.model.session.Session;
 import seedu.address.model.user.CardSrsData;
 import seedu.address.model.user.User;
