+++ src/test/java/seedu/address/logic/commands/quiz/QuizCommandTestUtil.java
@@ -6,7 +6,7 @@ import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.modelmanager.quiz.QuizModel;
+import seedu.address.model.modelmanager.QuizModel;
 
 /**
  * Contains helper methods for testing commands.
