+++ src/test/java/seedu/address/logic/parser/ParserUtilTest.java
@@ -10,7 +10,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.modelmanager.quiz.Quiz;
+import seedu.address.model.quiz.Quiz;
 
 public class ParserUtilTest {
     @Rule
