+++ src/test/java/seedu/address/logic/parser/StartCommandParserTest.java
@@ -8,7 +8,7 @@ import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailur
 import org.junit.Test;
 
 import seedu.address.logic.commands.quiz.QuizStartCommand;
-//import seedu.address.model.modelmanager.quiz.Quiz;
+//import seedu.address.model.quiz.Quiz;
 //import seedu.address.model.session.Session;
 
 public class StartCommandParserTest {
