+++ src/test/java/seedu/address/model/session/SessionTest.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 import org.junit.Test;
 
-import seedu.address.model.modelmanager.quiz.Quiz;
-import seedu.address.model.modelmanager.quiz.QuizCard;
+import seedu.address.model.quiz.Quiz;
+import seedu.address.model.quiz.QuizCard;
 import seedu.address.model.srscard.SrsCard;
 import seedu.address.testutil.Assert;
 import seedu.address.testutil.SessionBuilder;
